+++ src/main/java/kitchenpos/product/dto/ProductRequest.java
@@ -1,16 +1,15 @@
 package kitchenpos.product.dto;
 
-import java.math.BigDecimal;
 import kitchenpos.product.domain.ProductV2;
 
 public class ProductRequest {
     private String name;
-    private BigDecimal price;
+    private Long price;
 
     public ProductRequest() {
     }
 
-    public ProductRequest(String name, BigDecimal price) {
+    public ProductRequest(String name, Long price) {
         this.name = name;
         this.price = price;
     }
@@ -19,11 +18,11 @@ public class ProductRequest {
         return name;
     }
 
-    public BigDecimal getPrice() {
+    public Long getPrice() {
         return price;
     }
 
     public ProductV2 toProduct() {
-        return new ProductV2(null, this.name, this.price.longValue());
+        return new ProductV2(null, this.name, this.price);
     }
 }
