+++ src/main/java/kitchenpos/product/dto/ProductResponse.java
@@ -1,13 +1,11 @@
 package kitchenpos.product.dto;
 
-import java.math.BigDecimal;
-
 public class ProductResponse {
     private Long id;
     private String name;
-    private BigDecimal price;
+    private Long price;
 
-    public ProductResponse(Long id, String name, BigDecimal price) {
+    public ProductResponse(Long id, String name, Long price) {
         this.id = id;
         this.name = name;
         this.price = price;
@@ -21,7 +19,7 @@ public class ProductResponse {
         return name;
     }
 
-    public BigDecimal getPrice() {
+    public Long getPrice() {
         return price;
     }
 }
